% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_reduction.R
\name{atac_cca_label_transfer}
\alias{atac_cca_label_transfer}
\title{Transfer labels from a reference based on CCA results}
\usage{
atac_cca_label_transfer(
  ccv,
  reference_names,
  reference_labels,
  query_names,
  k = 20
)
}
\arguments{
\item{ccv}{CCV matrix, as generated by atac_lsi_cca()$ccv}

\item{reference_names}{row.names of ccv for cells from the reference set}

\item{reference_labels}{a vector of labels to transfer from the reference. Should have the same order and length as reference_names.}

\item{query_names}{row.names of ccv for cells from the query set}

\item{k}{Number of nearest-neighbors to use for label transfer. Default is 20}
}
\value{
A data.frame with 3 columns: barcodes (query_names) of query cells; reference_label, and reference_score.
}
\description{
Transfer labels from a reference based on CCA results
}
