% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_reduction.R
\name{atac_lsi}
\alias{atac_lsi}
\title{Perform Latent Semantic Indexing of a binary matrix}
\usage{
atac_lsi(atac_matrix, site_frequency_threshold = 0.03, seed = 3030)
}
\arguments{
\item{atac_matrix}{A binary, dgCMatrix of overlaps between cells (columns) and peaks (rows).}

\item{site_frequency_threshold}{A minimum threshold for the fraction of cells for which each peak must be positive to be retained.}

\item{seed}{A numeric value to use as a random seed. Default is 3030.}
}
\value{
A list with two objects: pcs, a VD matrix from singular value decomposition; and feature_loadings, a UD matrix from SVD.
}
\description{
This is a wrapper around two functions: atac_tf_idf() and atac_svd()
}
