% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_reduction.R
\name{atac_lsi_cca}
\alias{atac_lsi_cca}
\title{Use LSI results to project onto a common low-dimensionality space using CCA}
\usage{
atac_lsi_cca(atac_lsi_1, atac_lsi_2, seed = 3030)
}
\arguments{
\item{atac_lsi_1}{LSI representation of cells, i.e. as generated by atac_lsi()$pcs}

\item{atac_lsi_2}{LSI representation of cells, i.e. as generated by atac_lsi()$pcs}

\item{seed}{A numeric value to use as a random seed. Default is 3030.}
}
\value{
a list with two objects: ccv, a matrix of CC dimensions, and d from SVD
}
\description{
Use LSI results to project onto a common low-dimensionality space using CCA
}
