% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_reduction.R
\name{atac_lsi_projection}
\alias{atac_lsi_projection}
\title{Project a query TF-IDF dataset onto a target SVD space.}
\usage{
atac_lsi_projection(query_tf_idf, target_svd)
}
\arguments{
\item{query_tf_idf}{A matrix of TF-IDF results, e.g. as generated by atac_tf_idf()$tf_idf}

\item{target_svd}{SVD results for the target dataset, e.g. as generated by atac_svd() or atac_lsi()}
}
\description{
Both query_tf_idf and target_tf_idf must have common features.
}
