% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_reduction.R
\name{atac_svd}
\alias{atac_svd}
\title{Perform SVD on TF-IDF scaled values}
\usage{
atac_svd(tf_idf_mat, seed = 3030)
}
\arguments{
\item{tf_idf_mat}{A dgCMatrix of LogTF-IDF scaled values, e.g. as generated by atac_tf_idf()$tf_idf_mat}

\item{seed}{A numeric value to use as a random seed. Default is 3030.}
}
\value{
A list with two objects: pcs, a VD matrix from singular value decomposition; and feature_loadings, a UD matrix from SVD.
}
\description{
Perform SVD on TF-IDF scaled values
}
