% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_reduction.R
\name{atac_tf_idf}
\alias{atac_tf_idf}
\title{Perform LogTF-IDF on a sparse, binary matrix}
\usage{
atac_tf_idf(
  atac_matrix,
  site_frequency_threshold = 0.03,
  weight_threshold = site_frequency_threshold,
  idf = NULL
)
}
\arguments{
\item{atac_matrix}{A binary, dgCMatrix of overlaps between cells (columns) and peaks (rows).}

\item{site_frequency_threshold}{A minimum threshold for the fraction of cells for which each peak must be positive to be retained. Ignored if idf is supplied.}

\item{weight_threshold}{(optional) Threshold (>= site_frequency_threshold) to avoid over-weighting rare events. Default is equal to site_frequency_threshold.}

\item{idf}{(optional) A precomputed, named numeric vector of IDF weights to apply to the data. Default is NULL, which uses the atac_matrix to compute IDF.}
}
\value{
a sparse matrix of filtered and scaled count values
}
\description{
TF-IDF is Term Frequency-Inverse Document Frequency.
}
\details{
This function includes log scaling, as suggested by Andrew Hill.
}
