% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{binarize}
\alias{binarize}
\title{Binarize/convert all values in a sparse matrix to 1}
\usage{
binarize(mat)
}
\arguments{
\item{mat}{a sparse matrix object from the Matrix package, e.g. dgCMatrix.}
}
\value{
a sparse matrix object
}
\description{
Binarize/convert all values in a sparse matrix to 1
}
