% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{convert_fragments_gr}
\alias{convert_fragments_gr}
\title{Convert a list of fragment data.table to GenomicRanges}
\usage{
convert_fragments_gr(fragments, n_threads = 1)
}
\arguments{
\item{fragments}{A list of data.tables generated by read_10x_fragments()}

\item{n_threads}{A numeric value indicating how many threads to use for parallelization. Only works on UNIX-like systems. Default is 1.}
}
\value{
a list of GRanges objects
}
\description{
Convert a list of fragment data.table to GenomicRanges
}
