% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{convert_fragments_windows}
\alias{convert_fragments_windows}
\title{Convert a list of fragment data.tables to a sparse matrix of genomic window counts}
\usage{
convert_fragments_windows(
  fragments,
  window_size = 5000,
  genome = "hg38",
  n_threads = 1
)
}
\arguments{
\item{fragments}{A list of data.tables generated by read_10x_fragments()}

\item{window_size}{A numeric value specifying the size of windows to use as features. Default is 5000.}

\item{genome}{A character object specifying which genome to use. Should be one of "hg38","hg19","mm10", or "mm9". Default is "hg38".}

\item{n_threads}{A numeric value indicating how many threads to use for parallelization. Only works on UNIX-like systems. Default is 1.}
}
\value{
a list of GRanges objects
}
\description{
Convert a list of fragment data.tables to a sparse matrix of genomic window counts
}
