% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlaps.R
\name{count_frag_ol_ref}
\alias{count_frag_ol_ref}
\title{Count overlaps between each GRanges object in a list and a single target GRanges object (reference).}
\usage{
count_frag_ol_ref(
  query_fragments,
  target_GRanges,
  binarize = TRUE,
  sparse = FALSE,
  aggregate = FALSE,
  n_threads = 1
)
}
\arguments{
\item{query_fragments}{A GenomicRanges object or a list of GenomicRanges objects to use as a query}

\item{target_GRanges}{A single GenomicRanges object to use as a set of target regions}

\item{binarize}{A logical object indicating whether or not to binarize overlaps (count any number of overlaps between query and a given target region as 1). Default is TRUE.}

\item{sparse}{A logical object indicating whether the results should be a sparse matrix (TRUE) or a full matrix (FALSE). Default is TRUE.}

\item{aggregate}{A logical object indicating whether the results should be aggregated to a vector with the sum of counts for each query to all target regions. Default is FALSE.}

\item{n_threads}{A numeric object specifying the number of threads to use. Default is 1.}
}
\value{
If aggregate = TRUE, a vector of counts. If aggregate is FALSE and sparse is FALSE, a matrix object. If sparse is true, as dgCMatrix object.
}
\description{
Count overlaps between each GRanges object in a list and a single target GRanges object (reference).
}
