% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_diversity.R
\name{diversity_projection}
\alias{diversity_projection}
\title{Project library diversity using preseqR}
\usage{
diversity_projection(
  freq_mat,
  total_metrics,
  step_size = 1e+07,
  max_val = 5e+09,
  times = 100
)
}
\arguments{
\item{freq_mat}{a matrix containing the histogram of counts}

\item{total_metrics}{a data.frame of sequencing metrics}

\item{step_size}{a numeric value specifying how large of a step to use for projection. Default is 1e7}

\item{max_val}{a numeric value specifying the maximum extent of projection. Default is 5e9}

\item{times}{a numeric value specifying how many rounds of bootstrapping to use. Passed to preseqR.rSAC.bootstrap()}
}
\value{
a data.frame of projection results.
}
\description{
Project library diversity using preseqR
}
