% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{feature_jaccard}
\alias{feature_jaccard}
\title{Compute Jaccard distances between a vector (i.e. a single row) against all rows of a sparse matrix}
\usage{
feature_jaccard(
  query_vec,
  target_mat,
  axis = "row",
  type = "distance",
  direction = "positive"
)
}
\arguments{
\item{query_vec}{A binary vector object to use as a query.}

\item{target_mat}{A dgCMatrix to use as targets (will be treated as binary).}

\item{axis}{Either "row" or "col". Default is "row".}

\item{type}{distance Either "distance" or "similarity". distance is calculated as 1 - similarity. Default is "distance".}

\item{direction}{Either "positive" or "negative". "positive" uses intersection / union; "negative" uses setdiff / union. Default is "positive".}
}
\value{
a vector of Jaccard values between the query_vec and each row of the target_mat
}
\description{
Note: for efficiency, if axis = "row", will transpose and split the target_mat,
which will require available RAM equal to twice the size of target_mat.
}
\details{
If axis = "col", will only split, which requires available RAM equal to the size of target_mat.
}
