% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlaps.R
\name{filter_fragments}
\alias{filter_fragments}
\title{Filter a list of GenomicRanges objects against a single GenomicRanges object}
\usage{
filter_fragments(
  fragment_list,
  target_gr,
  mode = "remove",
  ignore_strand = TRUE,
  n_threads = 1
)
}
\arguments{
\item{fragment_list}{The list object containing GenomicRanges objects}

\item{target_gr}{The GenomicRanges object to use for filtering.}

\item{mode}{Whether to "remove" or "keep" overlaps. Default is "remove"}

\item{ignore_strand}{Logical, whether or not to ignore the strand of regions in the comparison}

\item{n_threads}{A numeric object specifying the number of threads to use. Default is 1.}
}
\value{
A list object containing filtered GenomicRanges objects.
}
\description{
Filter a list of GenomicRanges objects against a single GenomicRanges object
}
