% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlaps.R
\name{filter_gr}
\alias{filter_gr}
\title{Filter a GenomicRanges object against another GenomicRanges object}
\usage{
filter_gr(query_gr, target_gr, mode = "remove", ignore_strand = TRUE)
}
\arguments{
\item{target_gr}{The GenomicRanges object to use for filtering.}

\item{mode}{Whether to "remove" or "keep" overlaps. Default is "remove"}

\item{ignore_strand}{Logical, whether or not to ignore the strand of regions in the comparison}

\item{fragment_list}{The list object containing GenomicRanges objects}
}
\value{
A filtered GenomicRanges object.
}
\description{
Filter a GenomicRanges object against another GenomicRanges object
}
