% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{fragments_to_window_counts}
\alias{fragments_to_window_counts}
\title{Convert a data.table of fragment positions to a named vector of window counts}
\usage{
fragments_to_window_counts(fragment_dt, chrom_sizes, window_size = 5000)
}
\arguments{
\item{fragment_dt}{A data.table of fragment positions, e.g. as generated by read_10x_fragments()}

\item{chrom_sizes}{A data.frame of chromosome sizes, e.g. as generated by read_chrom_sizes()}

\item{window_size}{A numeric value specifying the size of windows to use as features. Default is 5000.}
}
\value{
A named numeric vector of reads per window. Names will be numeric window indexes.
}
\description{
Convert a data.table of fragment positions to a named vector of window counts
}
