% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archr.R
\name{get_archr_dgCMatrix}
\alias{get_archr_dgCMatrix}
\title{Get ArchR matrices in dgCMatrix format}
\usage{
get_archr_dgCMatrix(proj, target = "GeneScoreMatrix", ...)
}
\arguments{
\item{proj}{an ArchRProject object}

\item{target}{The name of a matrix in the ArchRProject. Options are: "GeneScoreMatrix", "PeakMatrix", "TileMatrix"}

\item{...}{Additional parameters passed to ArchR::getMatrixFromProject()}
}
\value{
a dgCMatrix object
}
\description{
Get ArchR matrices in dgCMatrix format
}
