% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archr.R
\name{get_archr_geneSet}
\alias{get_archr_geneSet}
\title{Pull the geneSet object from an ArchRProject}
\usage{
get_archr_geneSet(proj)
}
\arguments{
\item{proj}{an ArchRProject object}
}
\value{
a geneSet object
}
\description{
Pull the geneSet object from an ArchRProject
}
