% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archr.R
\name{get_archr_p2g_mat}
\alias{get_archr_p2g_mat}
\title{Get Peak-to-Gene correlations from an ArchRProject as a sparse matrix}
\usage{
get_archr_p2g_mat(proj, fdr_cut = 0.05, cor_cut = 0.3, cor_type = "any")
}
\arguments{
\item{proj}{an ArchRProject object}

\item{fdr_cut}{cutoff for p2g FDR. Default is 0.05}

\item{cor_cut}{cutoff for p2g correlation. Default is 0.3.}

\item{cor_type}{one of "pos" for positive correlations, "neg" for negative correlations, or "any" for either direction. Default is "any".}
}
\value{
a Matrix::dgCMatrix of Peak-to-Gene links with correlation scores as values
}
\description{
Get Peak-to-Gene correlations from an ArchRProject as a sparse matrix
}
