% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archr.R
\name{get_archr_peakAnno_mat}
\alias{get_archr_peakAnno_mat}
\title{Get a Peak Annotation sparse matrix from an ArchRProject}
\usage{
get_archr_peakAnno_mat(proj, peakAnnotation = "EncodeTFBS")
}
\arguments{
\item{proj}{an ArchRProject object}

\item{peakAnnotation}{The annotation to retrieve, e.g. "EncodeTFBS" or "Motif"}
}
\value{
a sparse matrix (usually logical) indicating which peaks match which annotations
}
\description{
Get a Peak Annotation sparse matrix from an ArchRProject
}
