% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archr.R
\name{get_archr_peakSet}
\alias{get_archr_peakSet}
\title{Pull the peakSet object from an ArchRProject}
\usage{
get_archr_peakSet(proj)
}
\arguments{
\item{proj}{an ArchRProject object}
}
\value{
a peakSet object
}
\description{
Pull the peakSet object from an ArchRProject
}
