% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archr.R
\name{link_mat_to_graph}
\alias{link_mat_to_graph}
\title{Convert a sparse linkage matrix to a graph object}
\usage{
link_mat_to_graph(link_mat)
}
\arguments{
\item{link_mat}{a matrix with sources as columns, targets as rows, and weights as values.}
}
\value{
an igraph object
}
\description{
Convert a sparse linkage matrix to a graph object
}
