% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archr.R
\name{link_peakAnno_to_gene}
\alias{link_peakAnno_to_gene}
\title{Link peakAnno to genes via Peak-to-Gene linkage}
\usage{
link_peakAnno_to_gene(anno_mat, p2g_mat)
}
\arguments{
\item{anno_mat}{a peakAnnotation matrix, as retrieved with get_archr_peakAnno_mat()}

\item{p2g_mat}{a Peak-to-Gene matrix, as retrieved with get_archr_p2g_mat()}
}
\value{
a Matrix::dgCMatrix of peakAnno-to-Gene links
}
\description{
This is matrix multiplication of the peakAnnotation matrix with the Peak-to-Gene matrix
}
