% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{load_reference}
\alias{load_reference}
\title{Load a stored reference dataset:}
\usage{
load_reference(
  name = c("altius", "gene_bodies", "great", "peaks", "tss"),
  genome = c("hg38", "hg19"),
  format = c("gr", "dt")
)
}
\arguments{
\item{name}{one of:
\itemize{
  \item{"altius"}{ENCODE/Altius index reference regions from Mueleman, et al. 2020 (ENCFF503GCK)}
  \item{"gene_bodies"}{Gene body regions from Ensembl (hg38 = v93; hg19 = v87)}
  \item{"great"}{Gene-adjacent regions computed a la GREAT}
  \item{"peaks"}{PBMC peaks identified in Lareau, et al. 2019 (GSE123577)}
  \item{"tss"}{TSS +/- 2kb regions from Ensembl (hg38 = v93; hg19 = v87)}
}}

\item{genome}{either "hg38" or "hg19".}

\item{format}{either "gr" for GenomicRanges or "dt" for data.table}
}
\value{
output based on the format parameter
}
\description{
Load a stored reference dataset:
}
