% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{max_column_cor}
\alias{max_column_cor}
\title{Compare two matrices, and find the column in the second matrix with maximum correlation to each column in the first matrix}
\usage{
max_column_cor(query_mat, target_mat, method = "pearson", n_threads = 1)
}
\arguments{
\item{query_mat}{A matrix object to use as a query.}

\item{target_mat}{A matrix object to use as a target.}

\item{method}{A character object specifying the correlation method to use. Passed to base R cor(). Default is "pearson".}

\item{n_threads}{A numeric object specifying the number of threads to use. Default is 1.}
}
\value{
a data.frame with stats for the top-2 best correlated columns in target for each column of query.
}
\description{
Compare two matrices, and find the column in the second matrix with maximum correlation to each column in the first matrix
}
