% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_chrom_sizes}
\alias{read_chrom_sizes}
\title{Read chromosome sizes for a genome and precompute offsets for windowing}
\usage{
read_chrom_sizes(genome = "hg38", window_size = 5000)
}
\arguments{
\item{genome}{A character object specifying which genome to use. Should be one of "hg38","hg19","mm10", or "mm9". Default is "hg38".}

\item{window_size}{A numeric value specifying the size of windows to use as features. Default is 5000.}
}
\value{
a data.frame with columns chr, size, n_windows, and offsets.
}
\description{
All chromosome sizes are sourced from the UCSC Genome Browser
}
