% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{sparse_jaccard}
\alias{sparse_jaccard}
\title{Compute Jaccard distances}
\usage{
sparse_jaccard(m)
}
\arguments{
\item{m}{a sparse matrix of KNN}
}
\value{
a sparse matrix of Jaccard distances.
}
\description{
Compute Jaccard distances
}
