% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential.R
\name{sparse_mat_chisq}
\alias{sparse_mat_chisq}
\title{Chi-squared tests for rows of a sparse matrix}
\usage{
sparse_mat_chisq(mat, fg_cols, bg_cols = NULL, rows = NULL, sort_on = "padj")
}
\arguments{
\item{mat}{A sparse matrix with observations as columns and features as rows.}

\item{fg_cols}{Column names for foreground columns}

\item{bg_cols}{Column names for background columns. If NULL (default) uses all non-fg_cols.}

\item{rows}{Row names to use for performing tests. If NULL (default), uses all rows}

\item{sort_on}{A character object specifying which column to use for sorting results. If NULL, will not sort. Default is "padj".}
}
\value{
a data.frame of Chi-squared test results.
}
\description{
Chi-squared tests for rows of a sparse matrix
}
